# docs/swagger.yaml
basePath: /api/v1
definitions:
  web.errorResponse:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
  web.response:
    properties:
      data: {}
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is an API for a clinic.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 'EBE3- FINAL: API CLINICA'
  version: "1.0"
paths:
  /odontologo:
    post:
      consumes:
      - application/json
      description: Create a new odontologo
      parameters:
      - description: auth token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Create a new odontologo
      tags:
      - Odontologo
  /odontologos:
    get:
      consumes:
      - application/json
      description: Get all odontologos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Get all odontologos
      tags:
      - Odontologo
  /odontologos/:id:
    delete:
      consumes:
      - application/json
      description: Delete odontologo by id
      parameters:
      - description: auth token
        in: header
        name: token
        required: true
        type: string
      - description: id del odontologo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Delete odontologo by id
      tags:
      - Odontologo
    get:
      consumes:
      - application/json
      description: Get odontologo by id
      parameters:
      - description: id del odontologo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Get odontologo by id
      tags:
      - Odontologo
    patch:
      consumes:
      - application/json
      description: Update all or some odontologo fields by id
      parameters:
      - description: auth token
        in: header
        name: token
        required: true
        type: string
      - description: id del odontologo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Complete or partial odontologo update by id
      tags:
      - Odontologo
    put:
      consumes:
      - application/json
      description: Update all odontologo fields by id
      parameters:
      - description: auth token
        in: header
        name: token
        required: true
        type: string
      - description: id del odontologo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Complete odontologo update by id
      tags:
      - Odontologo
  /paciente:
    post:
      consumes:
      - application/json
      description: Create a new paciente
      parameters:
      - description: auth token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Create a new paciente
      tags:
      - Paciente
  /pacientes:
    get:
      consumes:
      - application/json
      description: Get all pacientes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Get all pacientes
      tags:
      - Paciente
  /pacientes/:id:
    delete:
      consumes:
      - application/json
      description: Delete paciente by id
      parameters:
      - description: auth token
        in: header
        name: token
        required: true
        type: string
      - description: id del paciente
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Delete paciente by id
      tags:
      - Paciente
    get:
      consumes:
      - application/json
      description: Get paciente by id
      parameters:
      - description: id del paciente
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Get paciente by id
      tags:
      - Paciente
    patch:
      consumes:
      - application/json
      description: Update all or some paciente fields by id
      parameters:
      - description: auth token
        in: header
        name: token
        required: true
        type: string
      - description: id del paciente
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Complete or partial paciente update by id
      tags:
      - Paciente
    put:
      consumes:
      - application/json
      description: Update all paciente fields by id
      parameters:
      - description: auth token
        in: header
        name: token
        required: true
        type: string
      - description: id del paciente
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Complete paciente update by id
      tags:
      - Paciente
  /turnos:
    get:
      consumes:
      - application/json
      description: Get turno by DNI
      parameters:
      - description: dni del paciente
        in: query
        name: dni
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Get turno by DNI
      tags:
      - Turno
    post:
      consumes:
      - application/json
      description: Create a new turno
      parameters:
      - description: auth token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Create a new turno
      tags:
      - Turno
  /turnos/:id:
    delete:
      consumes:
      - application/json
      description: Delete turno by id
      parameters:
      - description: id del turno
        in: path
        name: id
        required: true
        type: integer
      - description: auth token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Delete turno by id
      tags:
      - Turno
    get:
      consumes:
      - application/json
      description: Get turno by id
      parameters:
      - description: id del turno
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Get turno by id
      tags:
      - Turno
    patch:
      consumes:
      - application/json
      description: Update all or some turno fields by id
      parameters:
      - description: id del turno
        in: path
        name: id
        required: true
        type: integer
      - description: auth token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Complete or partial turno update by id
      tags:
      - Turno
    put:
      consumes:
      - application/json
      description: Update all turno fields by id
      parameters:
      - description: auth token
        in: header
        name: token
        required: true
        type: string
      - description: id del odontologo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.errorResponse'
      summary: Complete turno update by id
      tags:
      - Turno
swagger: "2.0"